pyproject.toml
[project]
name = "testproject"
version = "0.1.0"
description = "A test project for verification"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "load-dotenv>=0.1.0",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.9.1",
    "structlog>=24.1.0",
    "PyYAML>=6.0",

] # Populated by user or later steps if needed

[project.scripts]
testproject = "test_project.main:main"

[project.optional-dependencies]
docs = [
    "sphinx>=8.1.2",
    "mkdocs-material>=9.5.0",
]

[tool.uv]
# These are the development tools installed by 'uv sync' if no specific extras are chosen,
# or if these are not specified in [project.optional-dependencies.dev] for 'uv pip install .[dev]'
# For this project template, these are the primary dev tools managed by nox.
dev-dependencies = [
    "nox>=2024.10.9",
    "pre-commit>=4.0.1",
    "ruff>=0.6.9",
    "pyright>=1.1.402",
    "pip-audit>=2.7.1",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "freezegun>=1.5.1",
    "pytest-mock>=3.14.0",
    "pytest-xdist>=3.5.0",
    "responses>=0.25.3",
    "diff-cover>=8.0.0",
    "commitizen>=3.20.0",
    "python-semantic-release>=8.10.0",
    "py-spy>=0.3.0",
    "scalene>=1.5.0",
    "hatch>=1.8.0",
    "pdm>=2.12.0"
    ]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
     "COM812", "COM819",
     "D100", "D203", "D213", "D300",
     "E111", "E114", "E117",
     "ISC001", "ISC002",
     "Q000", "Q001", "Q002", "Q003",
     "W191",
     "T",
     "D",
     "I001",  # Import sorting conflicts between tools - let ruff-format handle it
     "ANN401",  # Any types needed for structlog compatibility
     "FBT001", "FBT002",  # Boolean arguments acceptable for config functions
     "PLR0913",  # Many arguments acceptable for config functions
     "S110",  # Exception handling in optional imports is acceptable
     "BLE001",  # Broad exception handling needed for optional dependencies
     "PLC0415",  # Conditional imports needed for optional dependencies
 ]
 # Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "PLR2004", "TRY003", "EM101"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"


[tool.pyright]
include = ["src", "noxfile.py", "tests"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    ".venv",
    "venv",
    ".nox",
    "build",
    "dist",
]
venvPath = "."
venv = ".venv"
typeCheckingMode = "strict"
useLibraryCodeForTypes = true
reportMissingImports = true
reportMissingTypeStubs = false

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-v --cov=src/test_project -ra --strict-markers" # This will be replaced by setup.sh
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "integration: marks integration tests"
]

